requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
freetype-py==2.5.1
gunicorn==23.0.0
h11==0.14.0
hsluv==5.0.4
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
mutagen==1.47.0
numpy==2.2.3
openai==1.66.3
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-decouple==3.8
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
stripe==11.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vispy==0.14.3
watchdog==6.0.0
whitenoise==6.9.0
